Updated test coverage
