+++ src/access_nri_intake/catalog/translators.py
@@ -7,7 +7,7 @@ like the ACCESS-NRI catalog
 """
 
 from dataclasses import dataclass
-from functools import partial
+from functools import partial, wraps
 from typing import Callable
 
 import pandas as pd
@@ -15,7 +15,7 @@ import tlz
 from intake import DataSource
 
 from . import COLUMNS_WITH_ITERABLES
-from .utils import _to_tuple, trace_failure, tuplify_series
+from .utils import _to_tuple, tuplify_series
 
 # Note: important that when using @tuplify_series and @trace_failure decorators,
 # trace failure is the innermost decorator
@@ -37,6 +37,31 @@ FREQUENCY_TRANSLATIONS = {
     "yr": "1yr",
     "yrPt": "1yr",
 }
+
+
+def trace_failure(func: Callable) -> Callable:
+    """
+    Decorator that wraps a function and prints a message if it raises an exception
+    """
+
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        func_name = func.__name__
+        colname = func_name[1:].split("_")[0]
+        # Ensure the first argument is an instance of the class
+        if not isinstance(args[0], DefaultTranslator):
+            raise TypeError("Decorator can only be applied to class methods")
+
+        dispatch_key = getattr(args[0]._dispatch_keys, colname)
+
+        try:
+            return func(*args, **kwargs)
+        except KeyError as exc:
+            raise KeyError(
+                f"Unable to find {colname} column '{dispatch_key}' with translator '{args[0].__class__.__name__}'"
+            ) from exc
+
+    return wrapper
 
 
 class TranslatorError(Exception):
