+++ src/access_nri_intake/catalog/utils.py
@@ -1,4 +1,3 @@
-from functools import wraps
 from typing import Callable
 
 import pandas as pd
@@ -25,30 +24,5 @@ def tuplify_series(func: Callable) -> Callable:
     def wrapper(*args, **kwargs):
         series = func(*args, **kwargs)
         return _to_tuple(series)
-
-    return wrapper
-
-
-def trace_failure(func: Callable) -> Callable:
-    """
-    Decorator that wraps a function and prints a message if it raises an exception
-    """
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        func_name = func.__name__
-        colname = func_name[1:].split("_")[0]
-        # Ensure the first argument is an instance of the class
-        if not hasattr(args[0], "__class__"):
-            raise TypeError("Decorator can only be applied to class methods")
-
-        dispatch_key = getattr(args[0]._dispatch_keys, colname)
-
-        try:
-            return func(*args, **kwargs)
-        except KeyError as exc:
-            raise KeyError(
-                f"Unable to find {colname} column '{dispatch_key}' with translator '{args[0].__class__.__name__}'"
-            ) from exc
 
     return wrapper
